amf: ...oops

+++ ArcticFox.PolyType.Amf/AmfOptions.cs
@@ -29,14 +29,14 @@ namespace ArcticFox.PolyType.Amf
                 { typeof(string), new Amf0StringConverter() },
                 { typeof(float), new Amf0NumberConverter<float>() },
                 { typeof(double), new Amf0NumberConverter<double>() },
-                { typeof(int), new Amf0NumberConverter<sbyte>() },
-                { typeof(uint), new Amf0NumberConverter<byte>() },
-                { typeof(int), new Amf0NumberConverter<short>() },
-                { typeof(uint), new Amf0NumberConverter<ushort>() },
+                { typeof(sbyte), new Amf0NumberConverter<sbyte>() },
+                { typeof(byte), new Amf0NumberConverter<byte>() },
+                { typeof(short), new Amf0NumberConverter<short>() },
+                { typeof(ushort), new Amf0NumberConverter<ushort>() },
                 { typeof(int), new Amf0NumberConverter<int>() },
                 { typeof(uint), new Amf0NumberConverter<uint>() },
-                { typeof(int), new Amf0NumberConverter<long>() },
-                { typeof(uint), new Amf0NumberConverter<ulong>() },
+                { typeof(long), new Amf0NumberConverter<long>() },
+                { typeof(ulong), new Amf0NumberConverter<ulong>() },
                 { typeof(bool), new Amf0BoolConverter() }
             };
             m_amf0ConverterCaches = new MultiProviderTypeCache
